+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update base-image to java:8u402-3 to use doguctl v0.12.0 (#92)
+- Remove state since it is not used (#142)
 
 ## [v1.14.3-2] - 2024-06-06
 ### Fixed